requirements.txt
langchain
langchain-core
langchain-classic
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
wikipedia
langchain-text-splitters
pymupdf
langchain-openai
chromadb
faiss-cpu
langchain-ollama
langchain_chroma
sentence-transformers
langchain-openai
beautifulsoup4
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette